[FORMAL] Add another test case.
The example from issue is the code that we found in the issues of yosys thinking that it resolves our issue with submodule that have the same naming.
The problem_example shows the problem that we have with the tool.
Submodules are not checked by the tool and it only cares about the naming convention of the instances of the submodules